third_party/blink/web_tests/platform/generic/fast/forms/textfield-to-password-on-focus-expected.txt

Hit Tab key to focus input field.
Type some text.

Test for Bug 18900: Password field has focus but can't type text (i.rememberthemilk.com)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS field.value is 'pass'
PASS successfullyParsed is true

TEST COMPLETE
